backend/ai_detection/requirements.txt
# Core dependencies for deepfake detection
numpy
Pillow>=9.2.0
torch>=1.10.2
torchvision>=0.11.3

# Optional: For REST API server
# fastapi
# uvicorn
# pydantic